backend/requirements.txt
#frameworks
flask

#extensiones
flask-sqlalchemy
Flask-migrate
flask-marshmallow
flask-restful
flask-cors

#librerias
marshmallow-sqlalchemy
mysqlclient
python-dotenv
